requirements.txt
streamlit==1.29.0
pikepdf>=8.10.1,<9.0.0
python-docx==1.1.0
python-magic==0.4.27
google-generativeai==0.3.2
python-dotenv==1.0.0
cryptography==41.0.7
